feat: verify if exist notification

+++ src/modules/notifications/services/delete-notification/delete-notification.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { NotificationsRepository } from '../../repositories/notifications/notifications.repository';
+import { CustomerException } from '../../../../shared/errors/customerException';
 
 @Injectable()
 export class DeleteNotificationService {
@@ -8,6 +9,14 @@ export class DeleteNotificationService {
   ) {}
 
   public async execute(notificationId: string): Promise<void> {
+    const notification = await this.notificationsRepository.findByFilter({
+      notification_id: notificationId,
+    });
+
+    if (!notification) {
+      throw new CustomerException('Notification not found', 404);
+    }
+
     await this.notificationsRepository.deleteNotification(notificationId);
   }
 }
